+++ ci.nix
@@ -11,6 +11,7 @@
   nixpkgsVersions = {
     "R2003" = "nixpkgs-2003";
     "R2009" = "nixpkgs-2009";
+    "unstable" = "nixpkgs-unstable";
   };
   compilerNixNames = nixpkgsName: nixpkgs: builtins.mapAttrs (compiler-nix-name: runTests: {
     inherit (import ./default.nix { inherit checkMaterialization; }) nixpkgsArgs;
@@ -24,12 +25,14 @@
     # Update supported-ghc-versions.md to reflect any changes made here.
     {
       ghc865 = true;
+      ghc884 = false; # Just included because the native version is needed at eval time
     } // nixpkgs.lib.optionalAttrs (nixpkgsName == "R2009") {
-      ghc883 = false;
       ghc884 = true;
-      ghc8101 = false;
-      ghc8102 = true;
-      ghc810220201118 = false;
+      ghc8104 = true;
+    } // nixpkgs.lib.optionalAttrs (nixpkgsName == "unstable") {
+      ghc884 = true;
+      ghc8104 = true;
+      ghc810420210212 = false;
     });
   systems = nixpkgs: nixpkgs.lib.filterAttrs (_: v: builtins.elem v supportedSystems) {
     # I wanted to take these from 'lib.systems.examples', but apparently there isn't one for linux!
@@ -40,16 +43,19 @@
     # We need to use the actual nixpkgs version we're working with here, since the values
     # of 'lib.systems.examples' are not understood between all versions
     let lib = nixpkgs.lib;
-    in lib.optionalAttrs (nixpkgsName == "R2009" && (__elem compiler-nix-name ["ghc865" "ghc884"])) {
+    in lib.optionalAttrs (nixpkgsName == "unstable" && (__elem compiler-nix-name ["ghc865" "ghc884"])) {
     inherit (lib.systems.examples) ghcjs;
   } // lib.optionalAttrs (system == "x86_64-linux" && (
-         (nixpkgsName == "R2009" && __elem compiler-nix-name ["ghc8101" "ghc8102" "ghc810220201118"])
+         (nixpkgsName == "R2009" && __elem compiler-nix-name ["ghc8101" "ghc8102" "ghc8103" "ghc8104" "ghc810420210212"])
       || (nixpkgsName == "R2003" && __elem compiler-nix-name ["ghc865"]))) {
     # Windows cross compilation is currently broken on macOS
     inherit (lib.systems.examples) mingwW64;
-  } // lib.optionalAttrs (system == "x86_64-linux") {
+  } // lib.optionalAttrs (system == "x86_64-linux"
+      && !(nixpkgsName == "R2003" && compiler-nix-name == "ghc884")) {
     # Musl cross only works on linux
     # aarch64 cross only works on linux
+    # We also skip these for the R2003 was build of ghc884 (we only need the
+    # native so ifdLevel 1 includes compiler needed in ifdLevel2 eval)
     inherit (lib.systems.examples) musl64 aarch64-multiplatform;
   };
   isDisabled = d: