+++ overlays/bootstrap.nix
@@ -37,6 +37,13 @@ let
         else {
             compilerNixName = "ghc844";
         };
+    latestVer = {
+      "8.6" = "8.6.5";
+      "8.8" = "8.8.4";
+      "8.10" = "8.10.4";
+    };
+    traceWarnOld = v: x: __trace
+      "WARNING: ${x.src-spec.version} is out of date, consider using ${latestVer.${v}}." x;
 in {
   haskell-nix = prev.haskell-nix // {
     # Use this to disable the existing haskell infra structure for testing purposes
@@ -134,11 +141,15 @@ in {
                 ++ fromUntil "8.10" "8.10.2"   ./patches/ghc/67738db10010fd28a8e997b5c8f83ea591b88a0e.patch
                 ++ final.lib.optional (versionAtLeast "8.6.4" && versionLessThan "8.8") ./patches/ghc/ghc-no-system-linker.patch
 
-                ++ fromUntil "8.10.2" "8.12"   ./patches/ghc/MR3714-backported-to-8.10.2.patch
+                ++ fromUntil "8.10.2" "8.10.3" ./patches/ghc/MR3714-backported-to-8.10.2.patch
+
+                # See https://github.com/input-output-hk/haskell.nix/issues/1027
+                ++ final.lib.optional (versionAtLeast "8.10.3" && final.targetPlatform.isAarch64) ./patches/ghc/3434.patch
 
                 ++ from      "8.10.1"          ./patches/ghc/ghc-acrt-iob-func.patch
 
-                ++ fromUntil "8.10.1" "8.11"   ./patches/ghc/ghc-8.10-ubxt.patch
+                ++ fromUntil "8.10.1" "8.10.3" ./patches/ghc/ghc-8.10-ubxt.patch
+                ++ fromUntil "8.10.3" "8.11"   ./patches/ghc/ghc-8.10.3-ubxt.patch
                 ++ final.lib.optional (versionAtLeast "8.6.4") ./patches/ghc/Cabal-3886.patch
                 ;
         in ({
@@ -164,7 +175,7 @@ in {
                             ++ [ hsc2hs-align-conditionals-patch D5123-patch ]
                             ++ final.lib.optional final.stdenv.isDarwin ./patches/ghc/ghc-8.4.4-backport-dylib-command-size-limit.patch;
             };
-            ghc861 = final.callPackage ../compiler/ghc {
+            ghc861 = final.callPackage ../compiler/ghc (traceWarnOld "8.6" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc861; };
 
                 inherit sphinx installDeps;
@@ -183,8 +194,8 @@ in {
                 };
 
                 ghc-patches = [ D5123-patch ];
-            };
-            ghc862 = final.callPackage ../compiler/ghc {
+            });
+            ghc862 = final.callPackage ../compiler/ghc (traceWarnOld "8.6" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc862; };
 
                 inherit sphinx installDeps;
@@ -204,8 +215,8 @@ in {
 
                 ghc-patches = ghc-patches "8.6.2"
                             ++ [ D5123-patch ];
-            };
-            ghc863 = final.callPackage ../compiler/ghc {
+            });
+            ghc863 = final.callPackage ../compiler/ghc (traceWarnOld "8.6" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc863; };
 
                 inherit sphinx installDeps;
@@ -225,8 +236,8 @@ in {
 
                 ghc-patches = ghc-patches "8.6.3"
                             ++ [ D5123-patch ];
-            };
-            ghc864 = final.callPackage ../compiler/ghc {
+            });
+            ghc864 = final.callPackage ../compiler/ghc (traceWarnOld "8.6" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc864; };
 
                 inherit sphinx installDeps;
@@ -246,7 +257,7 @@ in {
 
                 ghc-patches = ghc-patches "8.6.4"
                             ++ [ D5123-patch ];
-            };
+            });
             ghc865 = final.callPackage ../compiler/ghc {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc865; };
 
@@ -268,7 +279,7 @@ in {
                 ghc-patches = ghc-patches "8.6.5"
                             ++ [ D5123-patch haddock-900-patch ];
             };
-            ghc881 = final.callPackage ../compiler/ghc {
+            ghc881 = final.callPackage ../compiler/ghc (traceWarnOld "8.8" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc881; };
 
                 inherit bootPkgs sphinx installDeps;
@@ -283,8 +294,8 @@ in {
                 };
 
                 ghc-patches = ghc-patches "8.8.1";
-            };
-            ghc882 = final.callPackage ../compiler/ghc {
+            });
+            ghc882 = final.callPackage ../compiler/ghc (traceWarnOld "8.8" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc882; };
 
                 inherit bootPkgs sphinx installDeps;
@@ -299,8 +310,8 @@ in {
                 };
 
                 ghc-patches = ghc-patches "8.8.2";
-            };
-            ghc883 = final.callPackage ../compiler/ghc {
+            });
+            ghc883 = final.callPackage ../compiler/ghc (traceWarnOld "8.8" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc883; };
 
                 inherit bootPkgs sphinx installDeps;
@@ -315,7 +326,7 @@ in {
                 };
 
                 ghc-patches = ghc-patches "8.8.3";
-            };
+            });
             ghc884 = final.callPackage ../compiler/ghc {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc884; };
 
@@ -332,7 +343,7 @@ in {
 
                 ghc-patches = ghc-patches "8.8.4";
             };
-            ghc8101 = final.callPackage ../compiler/ghc {
+            ghc8101 = final.callPackage ../compiler/ghc (traceWarnOld "8.10" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc8101; };
 
                 bootPkgs = bootPkgs // {
@@ -351,8 +362,8 @@ in {
                 };
 
                 ghc-patches = ghc-patches "8.10.1";
-            };
-            ghc8102 = final.callPackage ../compiler/ghc {
+            });
+            ghc8102 = final.callPackage ../compiler/ghc (traceWarnOld "8.10" {
                 extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc8102; };
 
                 bootPkgs = bootPkgs // {
@@ -371,10 +382,9 @@ in {
                 };
 
                 ghc-patches = ghc-patches "8.10.2";
-            };
-            # ghc 8.10.2 with patches needed by plutus
-            ghc810220201118 = final.callPackage ../compiler/ghc {
-                extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc810220201118; };
+            });
+            ghc8103 = final.callPackage ../compiler/ghc (traceWarnOld "8.10" {
+                extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc8103; };
 
                 bootPkgs = bootPkgs // {
                   # Not using 8.8 due to https://gitlab.haskell.org/ghc/ghc/-/issues/18143
@@ -386,16 +396,56 @@ in {
                 llvmPackages = final.llvmPackages_9;
 
                 src-spec = rec {
-                    version = "8.10.2";
+                    version = "8.10.3";
                     url = "https://downloads.haskell.org/~ghc/${version}/ghc-${version}-src.tar.xz";
-                    sha256 = "02w8n085bw38vyp694j0lfk5wcnwkdaj7hhp0saj71x74533lmww";
+                    sha256 = "0cdrdvs5qnqr93cr9zvrlfjv2xr671kjjghnsw4afa4hahcq7p6c";
                 };
 
-                ghc-patches = ghc-patches "8.10.2"
+                ghc-patches = ghc-patches "8.10.3";
+            });
+            ghc8104 = final.callPackage ../compiler/ghc {
+                extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc8104; };
+
+                bootPkgs = bootPkgs // {
+                  # Not using 8.8 due to https://gitlab.haskell.org/ghc/ghc/-/issues/18143
+                  ghc = final.buildPackages.buildPackages.haskell-nix.compiler.ghc865;
+                };
+                inherit sphinx installDeps;
+
+                buildLlvmPackages = final.buildPackages.llvmPackages_9;
+                llvmPackages = final.llvmPackages_9;
+
+                src-spec = rec {
+                    version = "8.10.4";
+                    url = "https://downloads.haskell.org/~ghc/${version}/ghc-${version}-src.tar.xz";
+                    sha256 = "03li4k10hxgyxcdyyz2092wx09spr1599hi0sxbh4m889qdqgbsj";
+                };
+
+                ghc-patches = ghc-patches "8.10.4";
+            };
+            # ghc 8.10.4 with patches needed by plutus
+            ghc810420210212 = final.callPackage ../compiler/ghc {
+                extra-passthru = { buildGHC = final.buildPackages.haskell-nix.compiler.ghc810420210212; };
+
+                bootPkgs = bootPkgs // {
+                  # Not using 8.8 due to https://gitlab.haskell.org/ghc/ghc/-/issues/18143
+                  ghc = final.buildPackages.buildPackages.haskell-nix.compiler.ghc865;
+                };
+                inherit sphinx installDeps;
+
+                buildLlvmPackages = final.buildPackages.llvmPackages_9;
+                llvmPackages = final.llvmPackages_9;
+
+                src-spec = rec {
+                    version = "8.10.4";
+                    url = "https://downloads.haskell.org/~ghc/${version}/ghc-${version}-src.tar.xz";
+                    sha256 = "03li4k10hxgyxcdyyz2092wx09spr1599hi0sxbh4m889qdqgbsj";
+                };
+                ghc-patches = ghc-patches "8.10.4"
                  ++ [ ./patches/ghc/core-field.patch ];
 
-                # Avoid clashes with normal ghc8102
-                ghc-version = "8.10.2.20201118";
+                # Avoid clashes with normal ghc8104
+                ghc-version = "8.10.4.20210212";
             };
         } // final.lib.optionalAttrs (final.targetPlatform.isGhcjs or false) (
          if final.hostPlatform.isGhcjs
@@ -538,6 +588,8 @@ in {
           final.haskell-nix.cabalProject ({
             name = "nix-tools";
             src = final.haskell-nix.sources.nix-tools;
+            # This is a handy way to use a local git clone of nix-tools when developing
+            # src = final.haskell-nix.haskellLib.cleanGit { name = "nix-tools"; src = ../../nix-tools; };
             index-state = final.haskell-nix.internalHackageIndexState;
             cabalProjectLocal = ''
               allow-newer: Cabal:base, cryptohash-sha512:base, haskeline:base
@@ -687,7 +739,7 @@ in {
               isHaskellNixBootCompiler = true;
             }
           )
-          (import ../compiler/old-ghc-nix { pkgs = final; });
+          (import sources.old-ghc-nix { pkgs = final; });
 
         packages = {
             # now that we have nix-tools and hpack, we can just