+++ overlays/hackage-quirks.nix
@@ -22,6 +22,11 @@ in { haskell-nix = prev.haskell-nix // {
         # Version of of cabal-install in hackage is broken for GHC 8.10.1
         (lib.optionalAttrs (version == "3.2.0.0") {
           packages.cabal-install.src = final.haskell-nix.sources.cabal-32 + "/cabal-install";
+          # Include patches needed for ghcjs
+          packages.Cabal.patches = [
+            ./patches/Cabal/Cabal-3.0.0.0-drop-pkg-db-check.diff
+            ./patches/Cabal/Cabal-3.0.0.0-no-final-checks.diff
+          ];
         })
       ];
     };
@@ -58,6 +63,14 @@ in { haskell-nix = prev.haskell-nix // {
           };
         })
       ];
+    };
+
+    haskell-language-server = {
+      cabalProject = ''
+        packages: .
+        allow-newer: haskell-language-server:ghcide
+        constraints: ghcide <0.7.4, hls-plugin-api <0.7.1.0, hls-retrie-plugin <0.1.1.1
+      '';
     };
 
   }."${name}" or {};