+++ test/index-state/default.nix
@@ -1,7 +1,7 @@
 # Test a package set
-{ stdenv, testSrc, tool, compiler-nix-name }:
+{ stdenv, lib, testSrc, tool, compiler-nix-name }:
 
-with stdenv.lib;
+with lib;
 
 let
   # The hackage-security 0.6.0.1 was uploaded at 2020-04-06T20:54:35Z
@@ -26,11 +26,11 @@ in
     buildCommand = ''
       if [[ "${version-before}" != "0.6.0.0" ]]; then
         echo 'Unexpected version ${version-before} (expected "0.6.0.0")'
-        exit 0
+        exit 1
       fi
       if [[ "${version-after}" != "0.6.0.1" ]]; then
         echo 'Unexpected version ${version-after} (expected "0.6.0.1")'
-        exit 0
+        exit 1
       fi
 
       touch $out